Add 'safe=False' flag to field mapper JsonResponses
Pylint recommends using JsonResponse instead of HttpResponse(blah,
content_type='application/json'). However, the data that the
field_mapper was passing to its JsonResponses in my tests was not a dict
and so we got a type error.

The Django docs state:
"Before the 5th edition of ECMAScript it was possible to poison the
JavaScript Array constructor. For this reason, Django does not allow
passing non-dict objects to the JsonResponse constructor by default.
However, most modern browsers implement EcmaScript 5 which removes this
attack vector. Therefore it is possible to disable this security
precaution."

Therefore, I feel reasonably confident putting 'safe=False' in the JsonResponse.